dribbble_mcp/requirements.txt
beautifulsoup4>=4.12.0
httpx>=0.27.0
lxml>=5.2.0
mcp>=1.0.0
Pillow>=10.3.0
playwright>=1.44.0
